Organize layer code
