+++ src/lib/browse/LayerControls.svelte
@@ -7,27 +7,27 @@
   } from "lib/common";
   import { CheckboxGroup } from "lib/govuk";
   import { interactiveMapLayersEnabled } from "stores";
-  import BusRoutesLayerControl from "./layers/BusRoutesLayerControl.svelte";
-  import CensusOutputAreaLayerControl from "./layers/CensusOutputAreaLayerControl.svelte";
-  import CombinedAuthoritiesLayerControl from "./layers/CombinedAuthoritiesLayerControl.svelte";
-  import CriticalIssuesLayerControl from "./layers/CriticalIssuesLayerControl.svelte";
-  import CrossingsLayerControl from "./layers/CrossingsLayerControl.svelte";
-  import CycleParkingLayerControl from "./layers/CycleParkingLayerControl.svelte";
-  import CyclePathsLayerControl from "./layers/CyclePathsLayerControl.svelte";
-  import HospitalsLayerControl from "./layers/HospitalsLayerControl.svelte";
-  import ImdLayerControl from "./layers/ImdLayerControl.svelte";
-  import LocalAuthorityDistrictsLayerControl from "./layers/LocalAuthorityDistrictsLayerControl.svelte";
-  import LocalPlanningAuthoritiesLayerControl from "./layers/LocalPlanningAuthoritiesLayerControl.svelte";
-  import MrnLayerControl from "./layers/MrnLayerControl.svelte";
-  import ParliamentaryConstituenciesLayerControl from "./layers/ParliamentaryConstituenciesLayerControl.svelte";
-  import PctLayerControl from "./layers/PctLayerControl.svelte";
-  import RailwayStationsLayerControl from "./layers/RailwayStationsLayerControl.svelte";
-  import RoadSpeedsLayerControl from "./layers/RoadSpeedsLayerControl.svelte";
-  import RoadWidthsLayerControl from "./layers/RoadWidthsLayerControl.svelte";
-  import SchoolsLayerControl from "./layers/SchoolsLayerControl.svelte";
-  import SportsSpacesLayerControl from "./layers/SportsSpacesLayerControl.svelte";
-  import VehicleCountsLayerControl from "./layers/VehicleCountsLayerControl.svelte";
-  import WardsLayerControl from "./layers/WardsLayerControl.svelte";
+  import CensusOutputAreaLayerControl from "./layers/areas/CensusOutputAreas.svelte";
+  import CombinedAuthoritiesLayerControl from "./layers/areas/CombinedAuthorities.svelte";
+  import ImdLayerControl from "./layers/areas/IMD.svelte";
+  import LocalAuthorityDistrictsLayerControl from "./layers/areas/LocalAuthorityDistricts.svelte";
+  import LocalPlanningAuthoritiesLayerControl from "./layers/areas/LocalPlanningAuthorities.svelte";
+  import ParliamentaryConstituenciesLayerControl from "./layers/areas/ParliamentaryConstituencies.svelte";
+  import WardsLayerControl from "./layers/areas/Wards.svelte";
+  import BusRoutesLayerControl from "./layers/lines/BusRoutes.svelte";
+  import CyclePathsLayerControl from "./layers/lines/CyclePaths.svelte";
+  import MrnLayerControl from "./layers/lines/MajorRoadNetwork.svelte";
+  import PctLayerControl from "./layers/lines/PCT.svelte";
+  import RoadSpeedsLayerControl from "./layers/lines/RoadSpeeds.svelte";
+  import RoadWidthsLayerControl from "./layers/lines/RoadWidths.svelte";
+  import CriticalIssuesLayerControl from "./layers/points/CriticalIssues.svelte";
+  import CrossingsLayerControl from "./layers/points/Crossings.svelte";
+  import CycleParkingLayerControl from "./layers/points/CycleParking.svelte";
+  import HospitalsLayerControl from "./layers/points/Hospitals.svelte";
+  import RailwayStationsLayerControl from "./layers/points/RailwayStations.svelte";
+  import SchoolsLayerControl from "./layers/points/Schools.svelte";
+  import SportsSpacesLayerControl from "./layers/points/SportsSpaces.svelte";
+  import VehicleCountsLayerControl from "./layers/points/VehicleCounts.svelte";
 
   export let style: string;
