+++ src/lib/browse/layers/areas/CensusOutputAreas.svelte
@@ -15,8 +15,8 @@
   } from "lib/maplibre";
   import type { MapGeoJSONFeature } from "maplibre-gl";
   import { map } from "stores";
-  import { colors } from "../colors";
-  import SequentialLegend from "./SequentialLegend.svelte";
+  import { colors } from "../../colors";
+  import SequentialLegend from "../SequentialLegend.svelte";
 
   let name = "census_output_areas";
   let outlineLayer = `${name}-outline`;
