+++ src/lib/browse/layers/areas/CombinedAuthorities.svelte
@@ -10,34 +10,28 @@
   import {
     hoveredToggle,
     overwriteLineLayer,
-    overwritePmtilesSource,
     overwritePolygonLayer,
+    overwriteSource,
   } from "lib/maplibre";
   import type { MapGeoJSONFeature } from "maplibre-gl";
   import { map } from "stores";
-  import { colors } from "../colors";
+  import { colors } from "../../colors";
 
-  let name = "wards";
-  let color = colors.wards;
+  let name = "combined_authorities";
+  let color = colors.combined_authorities;
   let outlineLayer = `${name}-outline`;
 
-  overwritePmtilesSource(
-    $map,
-    name,
-    `${publicResourceBaseUrl()}/v1/${name}.pmtiles`
-  );
+  overwriteSource($map, name, `${publicResourceBaseUrl()}/v1/${name}.geojson`);
 
   overwritePolygonLayer($map, {
     id: name,
     source: name,
-    sourceLayer: name,
     color,
     opacity: hoveredToggle(0.5, 0.0),
   });
   overwriteLineLayer($map, {
     id: outlineLayer,
     source: name,
-    sourceLayer: name,
     color,
     width: 2.5,
   });
@@ -59,23 +53,24 @@
   }
 
   function onClick(e: CustomEvent<MapGeoJSONFeature>) {
-    let name = encodeURIComponent(e.detail.properties.name);
-    // Help people find the councillor for this area
-    window.open(`https://www.google.com/search?q=${name}+councillor`, "_blank");
+    window.open(
+      `https://www.ons.gov.uk/visualisations/areas/${e.detail.properties.CAUTH22CD}`,
+      "_blank"
+    );
   }
 </script>
 
 <Checkbox id={name} bind:checked={show}>
   <ColorLegend {color} />
-  Wards
+  Combined authorities
   <span slot="right">
     <HelpButton>
       <p>
         Data from <ExternalLink
-          href="https://geoportal.statistics.gov.uk/datasets/ons::wards-may-2023-boundaries-uk-bgc/explore"
+          href="https://geoportal.statistics.gov.uk/datasets/ons::combined-authorities-december-2022-boundaries-en-buc/explore"
         >
           ONS Geography
-        </ExternalLink>, as of May 2023.
+        </ExternalLink>, as of December 2022.
       </p>
       <p>
         License: <ExternalLink
