+++ src/lib/browse/layers/lines/BusRoutes.svelte
@@ -14,7 +14,7 @@
   } from "lib/maplibre";
   import type { MapGeoJSONFeature } from "maplibre-gl";
   import { map } from "stores";
-  import { colors } from "../colors";
+  import { colors } from "../../colors";
 
   let name = "bus_routes";
