+++ src/lib/browse/layers/lines/PCT.svelte
@@ -14,8 +14,8 @@
   } from "lib/maplibre";
   import type { MapGeoJSONFeature } from "maplibre-gl";
   import { map } from "stores";
-  import { colors, denseLineWidth } from "../colors";
-  import SequentialLegend from "./SequentialLegend.svelte";
+  import { colors, denseLineWidth } from "../../colors";
+  import SequentialLegend from "../SequentialLegend.svelte";
 
   // TODO It'd be much simpler to have one source with both attributes
   let nameCommute = "pct_commute";
