+++ src/lib/browse/layers/points/Hospitals.svelte
