+++ src/lib/browse/layers/points/Schools.svelte
